Refactor gRPC client's tests
